Huge refactors because fixed #114 , separated and organized extensions
