+++ app/objectbox-models/default.json.bak
@@ -223,10 +223,6 @@
         {
           "id": "5:2686736484580824575",
           "name": "appearance"
-        },
-        {
-          "id": "6:211905571310986427",
-          "name": "parentID"
         }
       ],
       "relations": []
@@ -352,7 +348,8 @@
     990292594761795165,
     9203941479841948494,
     760325166941129440,
-    5171505801655906793
+    5171505801655906793,
+    211905571310986427
   ],
   "retiredRelationUids": [
     6347487514119236147,
