+++ app/src/main/java/uk/whitecrescent/waqti/backend/task/Label.kt
@@ -4,7 +4,7 @@ import io.objectbox.annotation.Entity
 import io.objectbox.annotation.Id
 import uk.whitecrescent.waqti.backend.Cacheable
 import uk.whitecrescent.waqti.backend.persistence.Caches
-import uk.whitecrescent.waqti.hash
+import uk.whitecrescent.waqti.extensions.hash
 
 // we've removed the nested labels feature for now as it may prove to be problematic
 @Entity
