+++ app/src/main/java/uk/whitecrescent/waqti/backend/task/Properties.kt
@@ -4,8 +4,8 @@ import com.google.gson.annotations.SerializedName
 import io.objectbox.converter.PropertyConverter
 import uk.whitecrescent.waqti.Duration
 import uk.whitecrescent.waqti.Time
-import uk.whitecrescent.waqti.fromJsonTo
-import uk.whitecrescent.waqti.toJson
+import uk.whitecrescent.waqti.extensions.fromJsonTo
+import uk.whitecrescent.waqti.extensions.toJson
 
 //region Properties
