+++ app/src/main/java/uk/whitecrescent/waqti/frontend/appearance/BoardAppearance.kt
@@ -6,8 +6,8 @@ import com.github.basshelal.unsplashpicker.data.UnsplashUrls
 import com.github.basshelal.unsplashpicker.data.UnsplashUser
 import com.google.gson.annotations.SerializedName
 import io.objectbox.converter.PropertyConverter
-import uk.whitecrescent.waqti.fromJsonTo
-import uk.whitecrescent.waqti.toJson
+import uk.whitecrescent.waqti.extensions.fromJsonTo
+import uk.whitecrescent.waqti.extensions.toJson
 
 class BoardAppearance {
