+++ app/src/main/java/uk/whitecrescent/waqti/frontend/fragments/create/CreateBoardFragment.kt
@@ -13,7 +13,11 @@ import org.jetbrains.anko.backgroundColor
 import uk.whitecrescent.waqti.R
 import uk.whitecrescent.waqti.backend.collections.Board
 import uk.whitecrescent.waqti.backend.persistence.Caches
-import uk.whitecrescent.waqti.commitTransaction
+import uk.whitecrescent.waqti.extensions.commitTransaction
+import uk.whitecrescent.waqti.extensions.getViewModel
+import uk.whitecrescent.waqti.extensions.invoke
+import uk.whitecrescent.waqti.extensions.requestFocusAndShowKeyboard
+import uk.whitecrescent.waqti.extensions.smoothScrollToEnd
 import uk.whitecrescent.waqti.frontend.CREATE_BOARD_FRAGMENT
 import uk.whitecrescent.waqti.frontend.FragmentNavigation
 import uk.whitecrescent.waqti.frontend.MainActivity
@@ -23,10 +27,6 @@ import uk.whitecrescent.waqti.frontend.appearance.BoardAppearance
 import uk.whitecrescent.waqti.frontend.appearance.WaqtiColor
 import uk.whitecrescent.waqti.frontend.fragments.parents.WaqtiCreateFragment
 import uk.whitecrescent.waqti.frontend.fragments.parents.WaqtiCreateFragmentViewModel
-import uk.whitecrescent.waqti.getViewModel
-import uk.whitecrescent.waqti.invoke
-import uk.whitecrescent.waqti.requestFocusAndShowKeyboard
-import uk.whitecrescent.waqti.smoothScrollToEnd
 
 class CreateBoardFragment : WaqtiCreateFragment<Board>() {
