+++ app/src/test/java/uk/whitecrescent/waqti/collections/BoardTests.kt
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
 import uk.whitecrescent.waqti.backend.persistence.Database
-import uk.whitecrescent.waqti.size
+import uk.whitecrescent.waqti.extensions.size
 import uk.whitecrescent.waqti.testBoardFullOfFullLists
 
 class BoardTests : BaseCollectionsTest() {
