+++ app/src/test/java/uk/whitecrescent/waqti/collections/TaskListTests.kt
@@ -6,8 +6,8 @@ import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
 import uk.whitecrescent.waqti.backend.persistence.Database
+import uk.whitecrescent.waqti.extensions.size
 import uk.whitecrescent.waqti.getTasks
-import uk.whitecrescent.waqti.size
 import uk.whitecrescent.waqti.testTask
 import uk.whitecrescent.waqti.testTaskListEmpty
 import uk.whitecrescent.waqti.testTaskListFull
