+++ app/src/test/java/uk/whitecrescent/waqti/task/Template.kt
@@ -24,8 +24,8 @@ import uk.whitecrescent.waqti.backend.task.OPTIONAL
 import uk.whitecrescent.waqti.backend.task.Priority
 import uk.whitecrescent.waqti.backend.task.Task
 import uk.whitecrescent.waqti.backend.task.Template
+import uk.whitecrescent.waqti.extensions.ids
 import uk.whitecrescent.waqti.hours
-import uk.whitecrescent.waqti.ids
 import uk.whitecrescent.waqti.minutes
 import uk.whitecrescent.waqti.now
 import uk.whitecrescent.waqti.tomorrow
