+++ build.gradle
@@ -2,7 +2,7 @@
 
 buildscript {
     ext.kotlin_version = '1.3.50'
-    ext.objectboxVersion = '2.3.4'
+    ext.objectboxVersion = '2.4.0'
     repositories {
         google()
         jcenter()
